backend/requirements.txt
django==4.1.7
sqlparse>=0.2.2
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
mysqlclient==2.1.1
regex==2022.10.31